+++ dart/lib/src/core/core/errors.g.dart
@@ -10,12 +10,14 @@ _$GameValue<G> _$$GameValueFromJson<G extends Game<Event>>(
         Map<String, dynamic> json) =>
     _$GameValue<G>(
       GameConverter<G>().fromJson(json['game'] as Map<String, dynamic>),
+      $type: json['runtimeType'] as String?,
     );
 
 Map<String, dynamic> _$$GameValueToJson<G extends Game<Event>>(
         _$GameValue<G> instance) =>
     <String, dynamic>{
       'game': GameConverter<G>().toJson(instance.game),
+      'runtimeType': instance.$type,
     };
 
 _$GameError<G> _$$GameErrorFromJson<G extends Game<Event>>(
@@ -23,6 +25,7 @@ _$GameError<G> _$$GameErrorFromJson<G extends Game<Event>>(
     _$GameError<G>(
       json['message'] as String,
       json['person'] as String,
+      $type: json['runtimeType'] as String?,
     );
 
 Map<String, dynamic> _$$GameErrorToJson<G extends Game<Event>>(
@@ -30,4 +33,5 @@ Map<String, dynamic> _$$GameErrorToJson<G extends Game<Event>>(
     <String, dynamic>{
       'message': instance.message,
       'person': instance.person,
+      'runtimeType': instance.$type,
     };